optimize achievements adapters
